fix: decode CAST(binary AS string) JVM-compatibly instead of unsafe reinterpret
Native CAST(binary AS string) built a Rust String from non-UTF-8 bytes via
from_utf8_unchecked, which is undefined behaviour (#4488), and an Arrow string
array holding invalid UTF-8 is unsound for the downstream native string kernels
that read it via StringArray::value.

Decode the bytes with decode_utf8_spark_lossy, which replaces ill-formed
sequences with U+FFFD exactly as the JVM's new String(bytes, UTF_8) does,
including the surrogate-range cases where Rust's from_utf8_lossy diverges. The
output is always valid UTF-8 and matches Spark's rendered result byte-for-byte,
so the cast stays Compatible (native, no opt-in). It diverges from Spark only
under byte-level round-trips such as CAST(CAST(x AS string) AS binary).

This decoder already existed in the native shuffle (#4521); move it into the
shared spark-expr utils module so shuffle and cast use one implementation.
Document the policy in the compatibility guide and the contributor guide.
